requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
faiss-cpu==1.12.0
fastapi==0.115.0
filelock==3.19.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.1.2
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pypdf==5.1.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==3.0.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.38.6
streamlit==1.48.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
